refactor!: reparameterize the penalty
Change the parameterization of the sorted l1 norm (and other upcoming
penalties) to take lambda as an argument in each of the member
functions, rather than to set it as a private member of the class. The
reason for this is that we need to work separately with lambda
elsewhere, so using `setLambda()` and `setAlpha()` methods along with
`getAlpha()` and `getLambda()` is just making us jump through
unneccesary hoops.