feat!: rename objective to loss
Rename Objective class to Loss, and change `setObjective()` to
`setLoss()`. The reason is that "objective" is sometwhat ambiguous since
it may refer to the term that include the penalty. Loss is somewhat less
ambiguous, but not perfectly so, so let's use that. Anothter term would
be "datafit", but I am reluctant to use it since it is not common.