perf(hybrid): use sparse collapsed cluster
Collapse the features of x for a cluster into a sparse representation
instead of storing it as a dense vector, which makes sparisty
meaningless when computing the gradient and hessians. Should improve
performance for (very) sparse x. Closes #123.

refactor: factor out some logic in hybrid updates

test(hybrid): add tests for hybrid method